Create new boilerplate code project
