+++ package.json
@@ -6,8 +6,7 @@
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "serve:ssr:boilerplate-testing": "node dist/boilerplate-testing/server/server.mjs"
+    "test": "ng test"
   },
   "type": "module",
   "private": false,
@@ -21,8 +20,7 @@
     "@angular/platform-browser-dynamic": "^19.2.4",
     "@angular/platform-server": "^19.2.4",
     "@angular/router": "^19.2.4",
-    "@angular/ssr": "^19.2.5",
-    "express": "^4.18.2",
+    "process": "^0.11.10",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
