+++ projects/boilerplate-testing-app/public/favicon.ico
